Execution/bridge_requirements.txt
python-dotenv
google-generativeai
slack-sdk
python-telegram-bot
requests
